fix(board-web-mvc): 사용자 관련 API를 따로 분리함
